+++ CHANGES.rst
@@ -4,6 +4,13 @@ Changelog
 1.0a13 (unreleased)
 -------------------
 
+New Features:
+
+- Add navigation and breadcrumbs as top-level services. Deprecate navigation
+  and breadcrumbs as components.
+  [timo]
+
+
 Bugfixes:
 
 - Add the title to the workflow history in the @workflow endpoint. This fixes
@@ -12,6 +19,7 @@ Bugfixes:
 
 - Don't fetch unnecessary PasswordResetTool in Plone 5.1
   [tomgross]
+
 
 1.0a12 (2017-04-03)
 -------------------